- Update shadow's driver
git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@6840 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: d0208cce965f289e349a2c5e4425e28fd4d5bb01
Former-commit-id: 96f7fd9e2a34b6786d51f5438854a9226bda3074

+++ src/drivers/shadow/src/Strategy.h
@@ -67,8 +67,8 @@ private:
 
     int				m_warnDamageLimit;
     int				m_dangerDamageLimit;
-    int             m_warnTyreLimit;
-    int             m_dangerTyreLimit;
+    double          m_warnTyreLimit;
+    double          m_dangerTyreLimit;
 
     bool            m_HasTYC;
